Refactored the lifecycle hooks
