+++ core/src/main/kotlin/io/specmatic/core/lifecycle/AfterLoadingStaticExamples.kt
@@ -1,0 +1,8 @@
+package io.specmatic.core.lifecycle
+
+import io.specmatic.core.Feature
+import io.specmatic.mock.ScenarioStub
+
+interface AfterLoadingStaticExamples {
+    fun call(examplesUsedFor: ExamplesUsedFor, examples: List<Pair<Feature, List<ScenarioStub>>>)
+}
