+++ core/src/main/kotlin/io/specmatic/core/lifecycle/ExamplesUsedFor.kt
@@ -1,0 +1,5 @@
+package io.specmatic.core.lifecycle
+
+enum class ExamplesUsedFor {
+    Test, Stub, Validation
+}
